myproject/requirement.txt
django
djangorestframework
pillow
#torch
torchvision
opencv-python
pytesseract
#numpy